Split traits and properties into seperate headers
