+++ executor/include/execution/property/blocking.hpp
@@ -1,0 +1,54 @@
+//
+// Created by Shrijit Singh on 14/06/20.
+//
+
+#pragma once
+
+#include <execution/property/base_property.hpp>
+
+class blocking_t {
+ public:
+  class always_t : public basic_executor_property<always_t, true, true> {
+   public:
+    template <class Executor>
+    friend Executor require(Executor &ex, const always_t &t) {
+      return ex.require(t);
+    }
+    template <class Executor>
+    friend bool query(const Executor &ex, const always_t &t) {
+      return std::is_same<always_t, decltype(ex.interface)>();
+    }
+  };
+
+  static constexpr always_t always{};
+
+  class never_t : public basic_executor_property<never_t, true, true> {
+   public:
+    template <class Executor>
+    friend Executor require(Executor &ex, const never_t &t) {
+      return ex.require(t);
+    }
+    template <class Executor>
+    friend bool query(const Executor &ex, const never_t &t) {
+      return std::is_same<never_t, decltype(ex.interface)>();
+    }
+  };
+
+  static constexpr never_t never{};
+
+  class possibly_t : public basic_executor_property<possibly_t, true, true> {
+   public:
+    template <class Executor>
+    friend Executor require(Executor &ex, const possibly_t &t) {
+      return ex.require(t);
+    }
+    template <class Executor>
+    friend bool query(const Executor &ex, const possibly_t &t) {
+      return std::is_same<never_t, decltype(ex.interface)>();
+    }
+  };
+
+  static constexpr possibly_t possibly{};
+};
+
+static constexpr blocking_t blocking;
